+++ storage/src/main/java/tech/pegasys/teku/storage/server/kvstore/dataaccess/V4FinalizedStateSnapshotStorageLogic.java
@@ -73,6 +73,9 @@ public class V4FinalizedStateSnapshotStorageLogic
       }
     }
 
+    @Override
+    public void commit() {}
+
     private void addFinalizedState(
         final KvStoreTransaction transaction,
         final SchemaFinalizedSnapshotState schema,