+++ storage/src/testFixtures/java/tech/pegasys/teku/storage/storageSystem/FileBackedStorageSystemBuilder.java
@@ -188,6 +188,7 @@ public class FileBackedStorageSystemBuilder {
         configDefault.withDatabaseDir(hotDir),
         storageMode,
         storeNonCanonicalBlocks,
+        10_000,
         spec);
   }