+++ teku/src/test/java/tech/pegasys/teku/cli/BeaconNodeCommandTest.java
@@ -485,7 +485,8 @@ public class BeaconNodeCommandTest extends AbstractBeaconNodeCommandTest {
                 b.eth1DepositContract(address)
                     .dataStorageMode(PRUNE)
                     .dataStorageFrequency(VersionedDatabaseFactory.DEFAULT_STORAGE_FREQUENCY)
-                    .dataStorageCreateDbVersion(DatabaseVersion.DEFAULT_VERSION))
+                    .dataStorageCreateDbVersion(DatabaseVersion.DEFAULT_VERSION)
+                    .maxKnownNodeCacheSize(100_000))
         .data(b -> b.dataBasePath(dataPath))
         .p2p(
             b ->
@@ -541,7 +542,8 @@ public class BeaconNodeCommandTest extends AbstractBeaconNodeCommandTest {
                     .metricsCategories(Set.of(BEACON, LIBP2P, NETWORK, EVENTBUS, JVM, PROCESS))
                     .metricsHostAllowlist(List.of("127.0.0.1", "localhost"))
                     .metricsEndpoint(null)
-                    .metricsPublicationInterval(60))
+                    .metricsPublicationInterval(60)
+                    .idleTimeoutSeconds(60))
         .interop(
             b ->
                 b.interopGenesisTime(1)