+++ CHANGELOG.md
@@ -18,7 +18,10 @@ For information on changes in released versions of Teku, see the [releases page]
 ## Unreleased Changes
 
 ### Additions and Improvements
+- Added `validator_local_validator_counts` metric to report number of local validators by current status.
+- Upgraded dependencies `io.netty`, `okhttp` and `io.vertex`.
 
 
 ### Bug Fixes
  - Fixed a possible crash on shutdown when using levelDb.
+ - Set an idle timeout for metrics connections, to clean up ports when no longer used